#34 パスによって表示するリストを変更できるよう変更
