+++ src/App.tsx
@@ -3,8 +3,8 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 
 import { AuthProvider } from './contexts/Auth';
 
-import PrivateRouter from './router/PrivateRouter';
-import PublicRouter from './router/PublicRouter';
+// import PrivateRouter from './router/PrivateRouter';
+// import PublicRouter from './router/PublicRouter';
 
 import Top from './components/Top';
 import Login from './components/Login';
@@ -19,15 +19,15 @@ const App: FC = () => (
         <Route path="/" exact>
           <Top />
         </Route>
-        <PublicRouter path="/login">
+        <Route path="/login">
           <Login />
-        </PublicRouter>
-        <PublicRouter path="/singup">
+        </Route>
+        <Route path="/singup">
           <Singup />
-        </PublicRouter>
-        <PrivateRouter path="/todo">
+        </Route>
+        <Route path="/todo">
           <Todo />
-        </PrivateRouter>
+        </Route>
         <Route>
           <NotFound />
         </Route>
